Fix trivial VC++ warnings in AndorLaserCombiner.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10830 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/AndorLaserCombiner/AndorLaserCombiner.cpp
@@ -539,7 +539,7 @@ int AndorLaserCombiner::OnPowerSetpoint(MM::PropertyBase* pProp, MM::ActionType 
    {
       pProp->Get(powerSetpoint);
 		LogMessage("to equipment: PowerSetpoint" + boost::lexical_cast<std::string, long>(Wavelength(il)) + "  = " + boost::lexical_cast<std::string,double>(powerSetpoint), true);
-		PowerSetpoint( il, (double)powerSetpoint);
+		PowerSetpoint( il, static_cast<float>(powerSetpoint));
 		if( openRequest_)
 			SetOpen();
